requirements.txt
absl-py==0.11.0
addict==2.4.0
albumentations==0.5.2
asgiref==3.2.10
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
autoflake==1.4
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.6.20
chardet==3.0.4
cli-exit-tools==1.1.8
click==7.1.2
cmake==3.18.2.post1
comtypes==1.1.7
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
distro==1.5.0
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
face-detection==0.2.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
image==1.5.32
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib-metadata==3.1.1
imutils==0.5.3
isort==5.6.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
Pillow==8.0.0
proglog==0.1.9
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-build==0.11.1
scikit-image==0.17.2
scipy==1.3.3
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu @ https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow_cpu-2.4.0-cp36-cp36m-manylinux2010_x86_64.whl
tensorflow-estimator==2.4.0
termcolor==1.1.0
tifffile==2020.9.3
toml==0.10.2
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.8.2%2Bcpu-cp36-cp36m-linux_x86_64.whl
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
